refactor(models): 删除原有的 models, 将代码迁移到 infrastructure 层的 DTO 当中
