--- app/models/__init__.py
@@ -1,2 +1,0 @@
-from .user import User
-from .comment import Comment
